requirements.txt
opencv-python
mediapipe
torch
torchvision
numpy
pyttsx3
Pillow
matplotlib
scikit-learn
tqdm
